refactor(web): reorganize components into domain directories
