+++ web/src/app/result/[id]/page.tsx
@@ -6,10 +6,10 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { animate } from "framer-motion";
 import { motion } from "framer-motion";
 import confetti from "canvas-confetti";
-import { CodePreview } from "@/components/code-preview";
-import { DeployStatus } from "@/components/deploy-status";
-import { DocViewer } from "@/components/doc-viewer";
-import { VibeScore } from "@/components/vibe-score";
+import { CodePreview } from "@/components/result/code-preview";
+import { DeployStatus } from "@/components/result/deploy-status";
+import { DocViewer } from "@/components/result/doc-viewer";
+import { VibeScore } from "@/components/result/vibe-score";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
