+++ web/src/components/brainstorm/index.ts
@@ -1,0 +1,1 @@
+export { BrainstormView } from "./brainstorm-view";
