+++ web/src/components/meeting/decision-gate.tsx
