+++ web/src/components/meeting/index.ts
@@ -1,0 +1,4 @@
+export { MeetingView } from "./meeting-view";
+export { CouncilMember } from "./council-member";
+export { CrossExam } from "./cross-exam";
+export { DecisionGate } from "./decision-gate";
