+++ web/src/components/result/index.ts
@@ -1,0 +1,4 @@
+export { VibeScore } from "./vibe-score";
+export { CodePreview } from "./code-preview";
+export { DocViewer } from "./doc-viewer";
+export { DeployStatus } from "./deploy-status";
