+++ web/src/components/result/vibe-score.tsx
