+++ web/src/components/shared/index.ts
@@ -1,0 +1,1 @@
+// Shared components will be added here (e.g., error-boundary in T9)
